Suppress config logs (#8124)
The idea was to make it visible that:
1. Specific setting is fetched from legacy Settings object instead of
   dynaconf
2. dynaconf validation failed - and should be fixed before we mandate
   dynaconf

However, it turned out to be misleading for team members, who thought
that logged exception signifies a real problem.

Suppress all logs from `robottelo.settings` for now. People working on
dynaconf integration can turn it back in during development. We can also
turn it back on once we are more confident in our dynaconf integration.

Also, make both main configuration module and facade write to same
logger namespace.