Using dynaconf formatting for repos
Adding a SATELLITE_VERSION field to robottelo config

The repo essentially supports a single satellite y-stream version on the
master branch at a given time, we just haven't programmatically
defaulted it anywhere, relying on SATELLITE_VERSION environment
variable.

This can break test collection if the envvar isn't set, so I see it fit
to actually establish, in config, what the 'active' version of Satellite
is on the master branch.

This would be changed for the 6.y.z branches