Removed parametrization causing errors in cli/test_repository.py
For test_negative_update_auth_url_with_special_characters and
test_negative_update_auth_url_too_long, the removed parametrization in
conjunction with the others caused only the first configuration of the
test to pass, while the following failed on errors during the setup of
the test.

That was because it tried to create a new repository with an already
existing name. Without the explicit parametrization, a random name
gets generated by default. Because the parametrization is therefore
unnecessary, I took the liberty of removing all of its instances in
the file.