Use dnf recommendation and add workaroud for bz_1983575
The issue related to hosts not getting some insights recomendations seems to be fixed, therefore
using the dnf recommendation for the test.
Also there is issue with syncing the recommendation for the first time on a fresh satellite.
Added a workaroud for it till the bz_1983575 is fixed. We can remove the work workaround
once the bugzilla is fixed.