Add a test for insights remediation feature (#8767)
* Add e2e insights recommendations sync and remediation test

* Add wait_for_tasks for insights sync task to finish

* Remove dnf recommendation

* Add wait_for_task and workaroud for search

* Resolve PR issues

* enable lab features

Insights remediation is still listead as a lab feature. Therefore we
need to enable lab_fetures setting for the tests to work. BZ_1976872
has been filed to track this. Once the BZ is fixed then we won't need
to enable lab_feature setting.

* Use dnf recommendation and add workaroud for bz_1983575

The issue related to hosts not getting some insights recomendations seems to be fixed, therefore
using the dnf recommendation for the test.
Also there is issue with syncing the recommendation for the first time on a fresh satellite.
Added a workaroud for it till the bz_1983575 is fixed. We can remove the work workaround
once the bugzilla is fixed.

* Add fixture to unset rh_cloud token after test execution

This commit adds a teardown fixture for test. The fixture will unset the
RH Cloud token setting at the end of the test execution.
It also adds workaround for search issue and sync issue.

* Resolve pr comments

* Add update_setting method and update job assertions

Add update_setting method in Satellite class and update the way of doing
assertion for job-invocation.

* print required data to debug test failure

* Pre-commit fix

* Fix assert statement