Isolate rhel support, params, and deploy
The content host yaml was trying to do far too much with the
increasingly abstracted test generation for rhel versions.

This creates a new supportability.yaml, which is not a template but a
direct definition of the supported versions of RHEL for the current
master branch of robottelo. This will change when backported to 6.10,
and in future master commits.

Remove handling of the workflow selection during test generation. The
fixture being indirectly parametrized already has a function that deals
with processing host arguments used for deployment.