Convert UI activation key tests to manifester (#10124)
* Convert UI activation key tests to manifester

This PR converts the ui/test_activationkey.py module to use Manifester
for tests that require manifests. To support this, the PR also adds four
function-scoped fixtures to pytest_fixtures/component/taxonomy.py.

* Create orgs with function-scoped fixtures

In the Joint Functional Teams meeting this morning, we discussed
Shweta's suggestion of replacing instances in which organizations are
created with `entities.Organization.create()` with function-scoped
manifest org fixtures. The consensus seemed to be that making this
change as part of the Manifester conversion process would be both
low-risk and low-effort, and that, while this change is technically
outside the scope of Manifester conversion, making it part of Manifester
conversion will likely allow us to complete it with less friction than a
separate effort in which individual component owners are responsible for
making the change in each of their components. This commit replaces
`entities.Organization.create()` with function-scoped manifest org
fixtures for all tests that use manifests in `ui/test_activationkey.py`.