Convert manifest-dependent API tests to manifester
This PR will convert all API tests that use manifests to obtain those
manifests from manifester. I am planning on pushing one commit per
module in an attempt to stay organized. The initial commit converts the
`api/test_contentview.py` module. Note that, in this module, the
TestContentViewRedHatOstreeContent class is decorated with
`@pytest.mark.skip_if_open("BZ:1625783")`, but this BZ is closed
WONTFIX. With the marker removed, these tests fail, but the failures are
unrelated to the use of manifester.