Convert manifest-dependent API tests to manifester (#10162)
* Convert manifest-dependent API tests to manifester

This PR will convert all API tests that use manifests to obtain those
manifests from manifester. I am planning on pushing one commit per
module in an attempt to stay organized. The initial commit converts the
`api/test_contentview.py` module. Note that, in this module, the
TestContentViewRedHatOstreeContent class is decorated with
`@pytest.mark.skip_if_open("BZ:1625783")`, but this BZ is closed
WONTFIX. With the marker removed, these tests fail, but the failures are
unrelated to the use of manifester.

* Resolve merge conflict

* Add debugger near failure point

* Remove debugger

* Convert API convert2rhel tests to manifester

This commit modifies the API convert2rhel tests to obtain their
manifests via manifester. Please note that these tests are currently
failing with the same error both locally and in CI.

* Convert API organization test to manifester

* Convert API product test to manifester

* Convert API report templates tests to manifester

* Resolve merge conflict

* Apply reformatting from black

* Convert API repository set tests to manifester

This commit converts the two tests in api/test_repository_set.py to
obtain manifests via manifester. Additionally, it simplifies the local
fixture stack by replacing two of the locally-defined fixtures with one
use of a function-scoped manifester fixture.

* Convert API errata tests to manifester

Note that these tests are failing in CI and locally. The failures do not
appear to be related to manifester.

* Convert API syncplan tests to manifester

* Convert API subscription tests to manifester

This commit converts all API subscription tests to manifester. Note that
this commit refactors some tests and adds an additional local fixture to
ensure compatibility with manifester. The tests should be functionally
equivalent, though.

* Resolve formatting issue

* Correct import issue

* Address feedback on subscription and contentview modules

* Consolidate variable assignment