[6.11.z] Cherry pick manifester fixtures (#10309)
* Cherry pick manifester fixtures to 6.11.z

The three PRs that introduced manifester fixtures failed to
auto-cherrypick in September and October. This PR manually cherry picks
the corresponding commits to the 6.11.z branch.

* Convert cli content view tests to manifester (#10098)

* Convert cli content view tests to manifester

This PR converts all tests that use manifests in cli/testcontentview.py
to obtain their manifests from manifester. To support this change, this
PR also adds four module-scoped fixtures (two that provide manifests and
two that provide organizations with manifests imported).

* Inherit module_org instead of module_target_sat in org fixtures

* Correct erroneous org return

* Cherry pick function-scoped manifester fixtures

This commit cherry picks the function-scoped manifester fixtures to
6.11.z.

* Retrigger GitHub actions

* Retrigger GitHub actions

* Fix requirements.txt for real this time