Rename /conf/manifest.yaml.example
Currently, the manifest config file template is named
manifest.yaml.example, which is inconsistent with the other template
filenames. Because of this, it was skipped by the config helper
script's filter for aligning symlinks. This PR renames the file to
manifest.yaml.template to improve naming consistency and to enable the
script to run on the file.