Modify entitlement manifester fixtures for 6.13
Satellite 6.13 changes the way that organizations and manifests interact
such that it is now required to explicitly disable simple content access
either at or after organization creation time in order to use
entitlement mode. This PR modifies the entitlement manifester fixtures
to account for this change. This should resolve some of the test
failures that we are seeing as a result of the new behavior. Please note that
Satellite will throw a 400 response if we attempt to change from SCA
mode to entitlement mode before a manifest has been uploaded to an org,
so it is necessary to upload the manifest in these fixtures before
running `org.sca_disable()`.