Convert UI errata tests to manifester
Previously, the UI errata tests used an internal `_org` method to create
an org, enable the `remote_execution_connection_by_ip` parameter, and
upload a manifest. This method was then returned by locally-defined
`org` and `module_org` fixtures to create a function-scoped and a
module-scoped org fixture with the parameter enabled on the org. This
approach was not compatible with the fixture scopes of the manifester
fixtures, so this commit removes the internal `_org` method and enables
the parameter separately in each locally-defined org fixture.