Revert to checking for type bytes in upload_manifest (#10554)
https://github.com/SatelliteQE/robottelo/pull/10458 introduced a change
to two conditional statements in the upload_manifest helper method in
satellite_mixins.py. The statements were changed from checking for a
manifest file of type `bytes` to one of `io.BytesIO`. This change
causes failures in tests using the manifester org fixtures, as
the `content` attribute of manifester manifests is a `bytes` object.
The content of cloned manifests is an `io.BytesIO` object, but the
conditionals in this method are designed around this difference, as the
two types of manifests need to be processed slightly differently here.
This PR reverts to checks for `bytes` objects rather than `io.BytesIO.`