Fix broker logging settings
We have added a new kwarg `file_level` in `setup_logzero` method that
broke robottelo as we did not supply this argument

The fix was to add that argument to the function call. Additionally the
logging settings has changed in a way that robottelo now reads the log
level from `logging.yaml` file.