Allow uploading cloned and manifester manifests (#10594)
After my recent PR reverting the change to the `isinstance` check in
`upload_manifest`, @Gauravtalreja1 pointed out to me that one of his PRs
was now failing on uploading manifests. That test was using cloned
manifests. This made me realize that, while the manifester conversions
are still in progress, we need to account for the possibility of either
type of manifest being used by a test. This PR modifies the `isinstance`
check to allow either type `bytes` or type `io.BytesIO`. This should
resolve both the errors that Gaurav encountered that caused him to make
the change initially and the errors that I encountered after that change
was merged.