[RFE] SAT-11254 Test Automation
test_positive_block_delete_key_in_use (ui and cli/test_contentcredentials.py)

-UI and CLI test automation
-BZ 2052904 : Blocker now in place when attempting to delete a content credential in use by a product or repository.

Both tests create a new gpg key, a new product with the key, a new repository associated with the product, and attempts to delete the key in use. The tests expect the gpg key to remain associated with the product and repo, and that the gpg, prod, and repo are unmodified.