Rewrite tests to have one assertion per test (#11086)
* Rewrite tests to have one assertion per test

It's a best practice to only have a single assertion per test. The
benefit is that all the other unrelated tests still run. This means that
you get a full report of all issues, and more precise reporting.

It does this by extracting the systemctl service status checks to a
separate, parametrized test. That is also rewritten to use systemctl
is-active instead of parsing the output of status. It still adds the
systemctl status output as context, which makes debugging easier.

The hammer ping command is then rewritten to collect all not-ok services
and assert that it's empty. This gives you a complete list of services
that aren't OK with their results, instead of just a single failed
service.

* Fixing docstring for test_positive_check_installer_service_running

---------

Co-authored-by: Griffin Sullivan <gsulliva@redhat.com>
(cherry picked from commit 28ce7ff2540ed90aabd7aa96900447e600049240)