Compare installer modules as sets
On the right hand side it is `dict_keys() - set()` which results in a
set. While you can compare dict_keys and set, the resulting diff isn't
very useful because pytest doesn't diff the elements.

This converts the left hand side into a set so it's comparing two sets.
Technically that's also correct because for us the order doesn't matter
either; it's only about the elements.