Avoid calling decode on str (#16214)
* Avoid calling decode on str

In tests we see this failure:

    AttributeError: 'str' object has no attribute 'decode'. Did you mean: 'encode'?

This implies that the strerr value is a str, which you can't decode. I
can't find the specific change, but this implies the result used to be a
byte string and is now decoded elsewhere.

* Update tests/foreman/destructive/test_rename.py

Co-authored-by: Peter Ondrejka <pondrejk@users.noreply.github.com>

* Update tests/foreman/destructive/test_rename.py

Co-authored-by: Evgeni Golov <evgeni@golov.de>

---------

Co-authored-by: Peter Ondrejka <pondrejk@users.noreply.github.com>
Co-authored-by: Jameer Pathan <21165044+jameerpathan111@users.noreply.github.com>
Co-authored-by: Evgeni Golov <evgeni@golov.de>
(cherry picked from commit 5c05d4fd72c0ad5472c0d8cc1a3702cce27f6f9b)