Qualify the container image names to localhost
We've seen some occasional issues where containers are exiting
prematurely. This may be to our config not qualifying the image name as
the one located on localhost.
With this change, we're being more explicit with that qualification.
I've also updated some of the changes in the config where our CI is ahead.