Add exception and remove 'upload_manifest=' arguments
This commit adds a `NoManifestProvidedError` to account for the
possibility of a user invoking the `upload_manifest` helper function
without providing a manifest argument. It also removes the
`upload_manifest=` argument from any tests and helpers that used it.
Almost all of our tests use `sca_manifest` or `sca_manifest_org`
fixtures, and our current plan is to standardize on that pattern moving
forward.