Convert Puppet class parameter upgrade scenario to new format
This PR converts the Puppet class parameter upgrade scenario to use the
SharedResource framework. It adds a fixture for a Puppet-enabled
Satellite, converts the pre-upgrade test to a fixture, removes a cleanup
helper function, adds a splay to ensure the `import_puppet` API endpoint
is not overloaded, and makes some additional small changes to ensure
that the test passes with the new framework.