[6.15.z] Remove fake manifest from Robottelo (#17907)
* Remove fake manifest from Robottelo (#17808)

* Remove fake manifest from Robottelo

This PR removes references to and uses of the fake manifest from all
Pytest fixtures, helper modules, utilities, and tests. Doing so will
enable Satellite QE to remove associated workflows and pipelines from
our AAP and Jenkins environments.

* Remove skip_if_not_set marks for fake manifest

* Add exception and remove 'upload_manifest=' arguments

This commit adds a `NoManifestProvidedError` to account for the
possibility of a user invoking the `upload_manifest` helper function
without providing a manifest argument. It also removes the
`upload_manifest=` argument from any tests and helpers that used it.
Almost all of our tests use `sca_manifest` or `sca_manifest_org`
fixtures, and our current plan is to standardize on that pattern moving
forward.

* Address incorrect fixture changes from merge conflict resolution