Check UI reaction to subscription_connection_enabled setting (#18259)
When subscription_connection_enabled setting is set to false the

    Automatic inventory upload switch
    Configure cloud connector button
    Sync all inventory status button

and these descriptions

To enable this reporting for all Foreman organizations, set Automatic inventory upload to on. The data will be reported automatically once per day.

To manually upload the data for a specific organization, select an organization and click Generate and upload report.

are hidden.