Convert content host scenario to new format
This PR converts the content host upgrade scenario to use the
shared_resource framework. It removes the existing class structure of
the module, converts the pre-upgrade test to a fixture, and moves
existing fixtures inherited by the old tests to the body of the
pre-upgrade fixture to ensure isolation from the main Robottelo fixture
stack.