Convert activation key upgrade scenario to new format (#16772)
* Convert activation key upgrade scenario to new format

* Address reviewer feedback

* Remove redundant deletes and update CODEOWNERS

This commit removes the steps to delete content view bits and custom
repos, as these steps are unrelated to the functionality being tested,
caused errors in a recent test run, and should not interfere with
parallel tests running on the same Satellite.

The commit also adds the new test to the Team Phoenix section of
`.github/CODEOWNERS`.