Convert Puppet scenario to new format (#18377)
* Convert Puppet scenario to new format

This PR converts the Puppet upgrade scenario to use the shared resource
framework. It adds a pre-upgrade fixture to ensure that the test can be
run concurrently with other upgrade scenario modules that require
Puppet-enabled Satellites and adds fixtures for the Satellite, Capsule,
and a Sat/Cap integration fixture that ensures Puppet is enabled on the
Capsule after it has already been set up for operation with the
Satellite.

* Address reviewer comments

* Address reviewer feedback

* Address reviewer feedback