[6.15.z] Convert user role scenario to new format (#18483)
Convert user role scenario to new format (#18390)

* Convert user role scenario to new format

This PR converts the user role upgrade scenario to use the
`shared_resource` framweork. It converts the pre-upgrade tests to
fixtures, removes the existing class structure for the non-stubbed
tests, and moves the non-stubbed tests to the top of the module above
the stubbed tests to improve readability.

* Address reviewer feedback