[6.17.z] Convert Puppet upgrade scenario to new format (#18587)
* Convert capsule upgrade scenarios to new format (#17343)

* Initial commit for capsule scenario conversion

* Get capsule upgrade tests passing with shared resources

* Use Capsule deploy workflow from conf/capsule.yaml

* Debug tests, incorporate framework changes from #17696

* Address heinous merge conflict

* Address fixture scope conflict

* Convert Puppet scenario to new format (#18377)

* Convert Puppet scenario to new format

This PR converts the Puppet upgrade scenario to use the shared resource
framework. It adds a pre-upgrade fixture to ensure that the test can be
run concurrently with other upgrade scenario modules that require
Puppet-enabled Satellites and adds fixtures for the Satellite, Capsule,
and a Sat/Cap integration fixture that ensures Puppet is enabled on the
Capsule after it has already been set up for operation with the
Satellite.

* Address reviewer comments

* Address reviewer feedback

* Address reviewer feedback

* Cherry pick parent PR commit and lint code