Update the card and disable caching support. (#18162)
Update the jira card in test. The first one is a duplicate and has been closed

Require below 2 changes -

Update the jira card in test. The first one is a duplicate and has been closed.
from SAT-23630 to SAT-31447

Others variable have three parameters, while firmware has four. Since they are used with zip, this mismatch is causing a conflict.