Convert usergroup upgrade scenario to new format (#18804)
This PR converts the usergroup upgrade scenario to use the
`SharedResource` framework. It converts the pre-upgrade test to a fixture,
adds a `SharedResource` fixture to manage the checkout of the Satellite,
and makes some minor changes to ensure that the test passes.

(cherry picked from commit b2898b000b2ae2ecc3829304144f76039e9a72e0)