Convert client upgrade scenario to new format
This PR converts the client upgrade scenario to use the `shared-resource`
framework. It converts the pre-upgrade test to a fixture, adds a new
`SharedResource` fixture for the Satellite, changes from a content host
fixture defined in the module to the global `rhel_contenthost` fixture,
and makes some additional minor changes to ensure that the tests pass.