Automate SAT-36449 and remove IOP parametrization (#19286)
This PR adds assertions to ensure that the errors detailed in SAT-36449
are not present. It also removes the parametrization for IOP from the RH
Cloud Insights end-to-end test, as the Recommendations and
Vulnerabilities pages are different enough between hosted and IOP that
we will be creating separate test modules and Airgun entities/views for
the on-prem scenario.

(cherry picked from commit cfc141a12e7a65348baae77c479a550ecc6ec574)