Add test for IOP inventory tags (#19338)
* Add test for IOP inventory tags

This PR adds a new test module for IOP API tests and a single test
checking for the present of inventory tags on a host reporting to an IOP
instance.

* Add assertion for Insights ID

* Mitigate race condition during teardown

Previously, I was seeing failures during teardown with this test, and I
believe I have traced them to a race condition caused by the
`module_target_sat_insights` fixture explicitly invoking a Broker
checkin. This resulted in there being no Satellite VM to check in when
the Pytest session reached its teardown phase. This commit removes the
explicit broker checkin.

* Invoke insights-client before getting host details