Add test for content view environment endpoint (#19581)
This PR adds a test for the content view environment endpoint. This test
ensures that results in responses from this endpoint contain a top-level
`id` field for the content view environment and that a search query to
the endpoint scoped to the `label` field is successful.

(cherry picked from commit 9541c182313fd71d7fbc996f05b29ead93b87fb5)