Block kubevirt tests on SAT-36027
Kubevirt as a compute resource is a tech preview feature, and
we do not currently have the infrastructure to test it reliably.
However, SAT-36027 tracks work on adding a compute resource for OCP-V
(which is downstream of Kubevirt) to Satellite. This PR blocks our
Kubevirt tests from running until this feature has been delivered, at
which time we should rewrite those tests for OCP-V.