Convert Lightspeed inventory sync from CLI to UI
The CLI test for syncing Lightpseed inventory status is failing due to
its reliance on the `foreman-rake rh_cloud_inventory:sync` rake task.
This command syncs the inventory for all organization on a Satellite and
does not respect options passed in to restrict the scope of the command
to a single org. As a result, in CI, when the test searches for an
inventory sync task on which to make assertions, it often erroneously
finds a sync task for an organization other than the one being tested.

Additionally, syncing the inventory via the `foreman-rake` CLI is not a
documented approach, and the `foreman-rake` CLI is typically not a
supported customer use case.

This PR converts the CLI inventory sync test to a UI test, which uses a
supported workflow and restricts the sync to the selected org by
default.