Add test for dangling transactions in Candlepin DB (#19854)
* Add test for dangling transactions in Candlepin DB

This PR adds a test for SAT-36234, a bug in which hung transactions in
Candlepin could eventually lead to a Postgres outage. This test is being
added to the subscription upgrade scenario, as starting Candlepin after
a database migration is one of the expected conditions for creating a
hung transaction prior to the fix for SAT-36234 being merged.

* Adjust pytest markers

(cherry picked from commit f10534ed9a5d7abe0cf4ddde53f417b797b1a34c)